Add sinon.addBehavior, use it to add the default behaviors
