+++ docs/release-source/release/migrating-to-2.0.md
@@ -53,3 +53,7 @@ The following utility functions are being marked as deprecated and are planned f
 ## `sandbox.useFakeXMLHttpRequest` no longer returns a "server"
 
 In Sinon 1.x, the sandbox' `useFakeXMLHttpRequest` was the same as it's `useFakeServer`. In 2.x, it maps directly to `sinon.useFakeXMLHttpRequest` (but with sandboxing). If you use `sandbox.useFakeXMLHttpRequest`, just replace it with `sandbox.useFakeServer`, and your tests should behave as they always did.
+
+## `sinon.behavior` is gone
+
+The `sinon.behavior` object is no longer exposed for random modification. However, there is a new mechanism in place aided to add new behavior to stubs, `sinon.addBehavior(name, fn)`, see the stub docs.
