+++ lib/sinon.js
@@ -64,3 +64,8 @@ exports.useFakeXMLHttpRequest = fakeXhr.useFakeXMLHttpRequest;
 exports.fakeServer = require("./sinon/util/fake_server");
 exports.fakeServerWithClock = require("./sinon/util/fake_server_with_clock");
 
+var behavior = require("./sinon/behavior");
+
+exports.addBehavior = function (name, fn) {
+    behavior.addBehavior(exports.stub, name, fn);
+};
