+++ lib/sinon/behavior.js
@@ -23,19 +23,6 @@ var nextTick = (function () {
     };
 })();
 
-function throwsException(error, message) {
-    if (typeof error === "string") {
-        this.exception = new Error(message || "");
-        this.exception.name = error;
-    } else if (!error) {
-        this.exception = new Error("Error");
-    } else {
-        this.exception = error;
-    }
-
-    return this.chain();
-}
-
 function getCallback(behavior, args) {
     var callArgAt = behavior.callArgAt;
 
@@ -114,6 +101,8 @@ var proto = {
     create: function create(stub) {
         var behavior = extend({}, proto);
         delete behavior.create;
+        delete behavior.addBehavior;
+        delete behavior.createBehavior;
         behavior.stub = stub;
 
         return behavior;
@@ -173,183 +162,6 @@ var proto = {
         );
     },
 
-    callsFake: function callsFake(fn) {
-        this.fakeFn = fn;
-        return this.chain();
-    },
-
-    callsArg: function callsArg(pos) {
-        if (typeof pos !== "number") {
-            throw new TypeError("argument index is not number");
-        }
-
-        this.callArgAt = pos;
-        this.callbackArguments = [];
-        this.callbackContext = undefined;
-        this.callArgProp = undefined;
-        this.callbackAsync = false;
-
-        return this.chain();
-    },
-
-    callsArgOn: function callsArgOn(pos, context) {
-        if (typeof pos !== "number") {
-            throw new TypeError("argument index is not number");
-        }
-
-        this.callArgAt = pos;
-        this.callbackArguments = [];
-        this.callbackContext = context;
-        this.callArgProp = undefined;
-        this.callbackAsync = false;
-
-        return this.chain();
-    },
-
-    callsArgWith: function callsArgWith(pos) {
-        if (typeof pos !== "number") {
-            throw new TypeError("argument index is not number");
-        }
-
-        this.callArgAt = pos;
-        this.callbackArguments = slice.call(arguments, 1);
-        this.callbackContext = undefined;
-        this.callArgProp = undefined;
-        this.callbackAsync = false;
-
-        return this.chain();
-    },
-
-    callsArgOnWith: function callsArgWith(pos, context) {
-        if (typeof pos !== "number") {
-            throw new TypeError("argument index is not number");
-        }
-
-        this.callArgAt = pos;
-        this.callbackArguments = slice.call(arguments, 2);
-        this.callbackContext = context;
-        this.callArgProp = undefined;
-        this.callbackAsync = false;
-
-        return this.chain();
-    },
-
-    yields: function () {
-        this.callArgAt = useLeftMostCallback;
-        this.callbackArguments = slice.call(arguments, 0);
-        this.callbackContext = undefined;
-        this.callArgProp = undefined;
-        this.callbackAsync = false;
-
-        return this.chain();
-    },
-
-    yieldsRight: function () {
-        this.callArgAt = useRightMostCallback;
-        this.callbackArguments = slice.call(arguments, 0);
-        this.callbackContext = undefined;
-        this.callArgProp = undefined;
-        this.callbackAsync = false;
-
-        return this.chain();
-    },
-
-    yieldsOn: function (context) {
-        this.callArgAt = useLeftMostCallback;
-        this.callbackArguments = slice.call(arguments, 1);
-        this.callbackContext = context;
-        this.callArgProp = undefined;
-        this.callbackAsync = false;
-
-        return this.chain();
-    },
-
-    yieldsTo: function (prop) {
-        this.callArgAt = useLeftMostCallback;
-        this.callbackArguments = slice.call(arguments, 1);
-        this.callbackContext = undefined;
-        this.callArgProp = prop;
-        this.callbackAsync = false;
-
-        return this.chain();
-    },
-
-    yieldsToOn: function (prop, context) {
-        this.callArgAt = useLeftMostCallback;
-        this.callbackArguments = slice.call(arguments, 2);
-        this.callbackContext = context;
-        this.callArgProp = prop;
-        this.callbackAsync = false;
-
-        return this.chain();
-    },
-
-    throws: throwsException,
-    throwsException: throwsException,
-
-    returns: function returns(value) {
-        this.returnValue = value;
-        this.resolve = false;
-        this.reject = false;
-        this.returnValueDefined = true;
-        this.exception = undefined;
-        this.fakeFn = undefined;
-
-        return this.chain();
-    },
-
-    returnsArg: function returnsArg(pos) {
-        if (typeof pos !== "number") {
-            throw new TypeError("argument index is not number");
-        }
-
-        this.returnArgAt = pos;
-
-        return this.chain();
-    },
-
-    returnsThis: function returnsThis() {
-        this.returnThis = true;
-
-        return this.chain();
-    },
-
-    resolves: function resolves(value) {
-        this.returnValue = value;
-        this.resolve = true;
-        this.reject = false;
-        this.returnValueDefined = true;
-        this.exception = undefined;
-        this.fakeFn = undefined;
-
-        return this.chain();
-    },
-
-    rejects: function rejects(error, message) {
-        var reason;
-        if (typeof error === "string") {
-            reason = new Error(message || "");
-            reason.name = error;
-        } else if (!error) {
-            reason = new Error("Error");
-        } else {
-            reason = error;
-        }
-        this.returnValue = reason;
-        this.resolve = false;
-        this.reject = true;
-        this.returnValueDefined = true;
-        this.exception = undefined;
-        this.fakeFn = undefined;
-
-        return this;
-    },
-
-    callThrough: function callThrough() {
-        this.callsThrough = true;
-        return this.chain();
-    },
-
     chain: function chain() {
         /**
          * "this" is stub when method is called directly on stub, e.g. stub.returns(123);
@@ -375,4 +187,23 @@ Object.keys(proto).forEach(function (method) {
     }
 });
 
+function createBehavior(behaviorMethod) {
+    return function () {
+        this.defaultBehavior = this.defaultBehavior || proto.create(this);
+        this.defaultBehavior[behaviorMethod].apply(this.defaultBehavior, arguments);
+        return this;
+    };
+}
+
+function addBehavior(stub, name, fn) {
+    proto[name] = function () {
+        fn.apply(this, arguments);
+        return this.stub || this;
+    };
+
+    stub[name] = createBehavior(name);
+}
+
+proto.addBehavior = addBehavior;
+proto.createBehavior = createBehavior;
 module.exports = proto;
