+++ lib/sinon/stub.js
@@ -1,6 +1,7 @@
 "use strict";
 
 var behavior = require("./behavior");
+var behaviors = require("./default-behaviors");
 var spy = require("./spy");
 var extend = require("./util/core/extend");
 var functionToString = require("./util/core/function-to-string");
@@ -127,24 +128,21 @@ var proto = {
     }
 };
 
-function createBehavior(behaviorMethod) {
-    return function () {
-        this.defaultBehavior = this.defaultBehavior || behavior.create(this);
-        this.defaultBehavior[behaviorMethod].apply(this.defaultBehavior, arguments);
-        return this;
-    };
-}
-
 Object.keys(behavior).forEach(function (method) {
     if (behavior.hasOwnProperty(method) &&
         !proto.hasOwnProperty(method) &&
         method !== "create" &&
         method !== "withArgs" &&
         method !== "invoke") {
-        proto[method] = createBehavior(method);
+        proto[method] = behavior.createBehavior(method);
+    }
+});
+
+Object.keys(behaviors).forEach(function (method) {
+    if (behaviors.hasOwnProperty(method) && !proto.hasOwnProperty(method)) {
+        behavior.addBehavior(stub, method, behaviors[method]);
     }
 });
 
 extend(stub, proto);
-
 module.exports = stub;
